Update README.md
This code is designed to analyze Boolean expression as a string, and to parse the string in order to extract data from it. It will then find the boolean index number of a the expression, and generate a truth table and PDNF. Then it will populate and filter the prime implicants that may be used in the argument, then search for a final solution.